Make Mac OS X installer build simpler
Checked in PackageIt.command and images/Installer3.png, which are the files needed to build the latest installer.  Modified .pro to copy them automatically to the build directory.  To release, update version numbers in .h file and in PackageIt.command file.  Build entire project (Release version).  Double click on PackageIt.command in the Build directory.  Test using the version in the BuildDir/Install.  This is identical to what users will install.