#441: formation string doesn't appear in SD window
Fixed.  The pointer to the graphicsText in the fixed version is not the same as the pointer to the graphicsText in the animated version.  Made two pointers, one for each graphicsText, and now it updates both of their text contents when the formation changeds  Also obsoleted the entire renderArea.cpp and renderArea.h files, which are no longer used.