#243: replay gain helper functions
Added asynchronous helper functions for replay gain.  These are called on Mac OS X only right now to generate a ReplayGain value for each MP3 song that is loaded.  Nothing is done with that value yet.  M4A not supported at this time (requires AACgain, which is a much larger subproject to bring in).