#243: Replay Gain now on in Mac OS X only
Replay Gain and Compressor are both available in the Mac OS X build now.  Replay Gain is still experimental at this time, but it seems to work pretty well.  If ReplayGain works well enough for most people, we can get rid of the Compressor, which I think is too confusing for most people.  ReplayGain is done via the DSP, because BASS_ATTRIB_VOL only goes from 0.0 to 1.0, despite what the documentation says.