#243: replayGain is now cached in DB
ReplayGain is saved once in the DB, and that value is used thereafter.  Can play a song, go into prefs, and toggle ReplayGain checkbox to hear the effect (this needs to continue to work -- it's useful to understand what ReplayGain does!).  Also added songsetting_attributes to the .pro file (it was missing).