#456: better BPM overriding
The code for BPM was pretty messed up.  Tempo override basically didn't work how I thought it worked, nor did it work the way the comments in the code indicated that it should.  I changed it to implement the way described in the Issue.  If Tempo Override is in effect, it overrides songs when loaded, and this value is persistent.