#462: fix rpath reference
I've made this mistake more than once.  @rpath on my machine DOES point at the Qt frameworks, while it does NOT on other machines.  So, it looks like it is working fine here, and it fails on other people's machines.  This commit changes to using @executable_path instead, which should work on all platforms.