#497: unknown types are now patter
If I stick a bunch of files into a directory at the main music directory level, they will default to Patter type, and therefore they will default to LOOP ON.  This is quite surprising if you don't notice it, because the singing call will loop right near the end.  Oops!  Now unknown types default to NO LOOP (they are are not singing calls, so you don't get the colored bars).