#501: Reveal in Finder for Lyrics
Right-click will ask the Finder to pull up a window, with the lyrics file highlighted.  You can then click on the Share icon at the top (Apple Mail), or if you have the special "Send to Thunderbird" workflow installed, use Right Click > Services > Send to Thunderbird, and the file will be an attachment to an email, so you can share the file with your friends.