Initial working audioDecoder
This audioDecoder is based on one of the Examples, and it works.  It only loads a single song right now, but it does so in about 250ms.  The result is a QByteArray that we should be able to do BPM, EQ, and PLAY on.  It will properly delete the memory when a new file is loaded via setSource.  This is intended to be a singleton.  Tested on M1 Mac.