WIP: code cleanup, fixed Play/Pause bug
Play/Pause button now changes state correctly (required plumbing currentStreamState through to the player.  Remove the other 2 implementations of MP3 decode.  Now using the one in AudioDecoder, which is also a player.  flexible_audio is now a wrapper around that, a wrapper that is compatible with the bass_audio subsystem that the new system replaces.