Bug: fixed QThread crash on EQ slider moves
I was deleting the filter out from under the player whenever the EQ changed, and it was not happy.  Now, I only make a new filter just before the player needs it.  It's at that precise point that I know that the filter is both NOT in use, and it needs to be re-initialized (because EQ has changed).