WIP: initial RubberBand attempt
The pitch slider is hooked up, the tempo slider is not.  There's some sort of conflict between samples needed by the speaker and samples needed by the RubberBandStretcher.  Playing a tone works on the Intel SquareDesk without interruptions in the audio.  But right now, the tone played on the M1 SquareDesk is kinda correct, but has periodic audio problems.  Also, when the tempo is greater than 1.0 (slower that realtime), the RubberBandStretcher fills up with samples.  So, some problem with the handoff.  But, I wanted to check this in anyway, as an initial cut at it, while I work on it more.