WIP: tempo changing
RubberBand kinda sorta works.  It appears to run out of data, though, if the tempo is set to higher than the nominal tempo.   It does not recover in this case, until I lower the tempo for a while.  Maybe RubberBand is falling behind realtime?  More likely is that I have a bug here.  One step at a time.