#599: changing sound device in Sound Preferences now plays audio
There's still a Qt bug here, in that I am notified only when the audio output device list changes, NOT when the default audio device changes.  I'd have to poll -- maybe someday I will.  In the mean time, I check the default audio output device whenever PLAY is pressed, and then (and only then) do we change  the QAudioSink to a new one.   This will probably solve the problem for most people.  Plugging in headphones while the player is STOPPED should work as expected.