Initial work for libtidy
Currently commented out, because the latest version of libtidy is destroying the formatting of the lyrics file when it is used.  But, the latest version of libtidy is indeed being put into the build now.  Uncomment the QMAKE lines and #undef NOTIDYYET to put it back in for debugging this.