#662: added info to Export Song History
Now exports filename (because just the song name is ambiguous), pitch, tempo, and last_cuesheet.  Note that the latter 3 fields are the CURRENT values, not actually what was played at the time.  This is a quick and dirty hack that works only if I don't change those between playback time and Export Play History time.