Clear out com.apple.quarantine from all files
I think this is why Regina was having trouble running the 0.9.10 version.  One of the files, sd_doc.pdf, was listed as quarantined.  This change double checks to clear any extended attributes, for all files in the SquareDesk.app .  The theory is that on Regina's side, it will install without turning on the com.apple.quarantine bit.  Manually running xattr -cr on SquareDesk_*.app fixed the problem for Regina, but this is painful.