#676: cuesheet.css is now embedded in each cuesheet
Standalone cuesheet.css is now essentially obsolete.  Everything is styled to the internal cuesheet now, that is loaded into the QTextEdit at startup.  That cuesheet is now written into every cuesheet HTML file.  That means that every cuesheet is now standalone -- it can be viewed in a browser, or sent around to a friend.  I think it will probably work fine inside sqview, although I have not tested that.