Update M1 build to Qt6.3.1
Had to include mytablewidget.h.  Not sure why this succeeded before and it was just detected now (with QtCreator 8), but it seems to be fixed by this change.  Now the M1 and X86 builds are both on Qt6.3.1, so I think that both builds can now use PackageIt_M1.command without changes.