#721: Super Clean lyrics HTML files made in other programs
MS Word and OpenOffice produce pretty terrible HTML files, but music producers often use those programs to make cuesheets for distribution with their music.  Now we have a way to get rid of most of the mess.  Unlock the Lyrics file for editing, use Cmd-A to select the whole thing, and click Clear Formatting.  The entire <HEAD> section will be deleted, and all HTML tags will be removed (except for <BR> which means that line breaks are preserved).  Sweet.