#736: Adjust spacing in EQ section of the Music Player tab
sharmanr is reporting that he needs to make the height of the containing box smaller to make the B/M/T not get cut off.  I'm not sure why I'm seeing something different, although it could be because I'm on Qt6.4.1 now.    Changing the height of the EQ bottom spacer from 16 to 1px makes it much better for me.