#725: If playlist was made from scratch and Save As'd, remember that
Next time SquareDesk starts up, it will load that playlist.  It did this already for playlists that were explicitly loaded, but NOT for playlists that were just created from scratch.  This commit fixes that.  It also fixes a problem whereby the Save Playlist '<playlist name>' wasn't setting the playlist name, for newly created playlists.