#736: BMT gets cut off at small zoom levels
It turns out that it's the QGroupBox that is preventing proper resizing behavior.  For now, I've removed the QGroupBox from the hierarchy.  We lose the "3-band EQ" text, but we gain back auto resizing.  I'd rather have it resize properly at all zoom levels, if I am forced to choose.