#722: Part 1 of auto formatting for lyrics
This is intended to be used after Unlock for Editing, Select All, Lyrics > Clear Formatting, or when pasting into an empty document from a text editor (or copied from a PDF file).  First, select text to change (usually ALL text with Cmd-A).  If a typical header phrase (OPENER, MIDDLE BREAK, CLOSER, TAG, FIGURE n) is at the beginning of a line, it will automatically be highlighted as HEADER type.  The term "BREAK" at the beginning of a line will automatically be changed to "MIDDLE BREAK", and highlighted as HEADER type.  And, the following will be automatically capitalized (HEADS, SIDES, BOYS, GIRLS), as is typical in calls.