#729: User-supplied default templates for cuesheets
If a template override file is present, will be used in place of the default ones that are embedded into the Resources folder of the app itself (Mac OS).  Note that a user-defined template override must be put into the correct directory:
lyrics -->  <musicDir>/lyrics/lyrics.template.html
patter --> <musicDir>/patter/patter.template.html
By putting it here instead of the <musicDir>/reference folder, we don't have to filter it out from the Reference tabs, and it's extensible to whatever directories the user decides to create.  I think it's somewhat more intuitive, too.