#630: Remove the string edit hack
FormatTitlePlusTags() now takes an optional fourth argument.  If not specified, it works like it did before (but there are currently no call sites that use this.  If specified, the title part of the string will be wrapped in a span that specifies its foreground color.  No magic string replacements at each call site needed now.