#752: fix switching between focus in Current Sequence and SDLineEdit
Now when you click (NOT double click) on an item in the Current Sequence pane, we'll render it.  This will also Select that item (highlights in BLUE).  When clicking back to the SDLineEdit widget, the selection will be cleared, and we'll render the last call in the CurrentSequence window, as expected "back to normal call entering mode".