#752: Initial Frames UX prototype
The content of the frames is static right now, but the look and feel is kinda working now.  The flow is left-to-right from Frame List on the left to Current Frame to picture (when in Arranger mode).   Hitting a F-key will swap in that frame to the Current Frame position, at the current call.