#752: initial SD editing buttons
Three buttons, that appear at various times.  NEW will make a new blank sequence, and append it to the Current Frame file, and open up the UI for editing.  UNLOCK will allow editing of the Current Sequence in the Current Frame file, and will open up the UI for editing.  SAVE has many variants (to be implemented), including MOVE TO FRAME and COPY TO FRAME, both of which will exit edit mode.  Currently, if fake data is used, these buttons will not appear, so as not to confuse current users.