#752: Persist the frameCurSeq values
As the user presses F11/F12 to move around within a frame, we keep track of that internally for displaying the frame (either in the sidebar, or in the Current Sequence pane).  This commit persists that value at quit() time, so the next time the user comes in, the frameCurSeq values are remembered.  If the persist file doesn't exist, it is created, with an initial value of 1 for all files.