#752: LEFT/RIGHT arrow keys act like F11/F12
If on the SD page, Left/Right Arrows will go to Previous/Next sequence in the current frame, as expected.  NOTE:  If you have the default keyboard shortcuts setup, Left/Right are assigned to moving around within a song.  In that case, two alternatives:  1. clear out those shortcuts in the HotKeys tab in preferences, or 2. use Cmd-LEFT and Cmd_RIGHT which work even when the shortcuts are set to LEFT and RIGHT arrows.