#752: REC= and AUTHOR= records now contain data
REC contains a userID in the top 5 decimal digits, and a SequenceID in the lower 5 digits.  The SequenceID is incremented whenever the sequence is created or altered.  The AUTHOR field contains the userID (same as used in the REC) -- this will be changed to an author string at a later date.