#752: Add authorID
If there isn't already a .squaredesk/metadata.csv file, the authorID is set to the same as userID (it's a number).  If you want it to be a real string, you'll have to edit this by hand right now, e.g. "authorID,Mike Pogue".  Then, that string will be used for the #AUTHOR=<authorID># records in saved sequences.