#752: SD Level is now persistent
The default if you've never selected a level is "Plus".  But, when you select a level, that selection will persist to the next time you open SquareDesk.  Callers with a mixture of levels should set the SD level to the highest level they expect to use (otherwise, at SD startup time, a C1 sequence in the default Current Sequence window might not load properly if Plus is the default level).