#752: edit sequence indicator, fix seq numbering, fix at sign at EOF
This commit adds an edit sequence indicator, which is triggered either on NEW or UNLOCK.  However, NEW will create a sequence, but now will NOT write it to the file.  That will be done at SAVE time.  If the app is quit between NEW and SAVE, nothing is written to the file.  AND, null sequences are written as "just as you are", which makes them both viewable and editable.