#758: speed up Move To Top/Bottom of Playlist
It was, in fact, not turning off sorting.  This was only a problem for larger playlists, but one user had a playlist of 115 items, so moving to top/bottom was taking on the order of 2-3 seconds.  Temporarily turning off sorting using start/stopLongSongTableOperation() dropped the time in my testing (playlist 30 items) from ~2 seconds to ~200ms.