#752: when NEW is clicked, set focus to the SDInput field
When NEW is clicked, we want to start typing calls ASAP, without having to manually click in the SD Input field to give it focus.  This is a bit tricky, because SD is asynchronous, so we don't get control until later.  We already have a timer to do this (I don't know of any other way), so if the timer goes off, SD is done, and we can (if the SD Input field is visible, AND we just did a NEW operation) set focus to the SDInput field.