#752: When G or B is used, DO go to the next sequence
I had it this way before, and thought it might work better NOT moving to the next sequence.  But in actual use, my expectation was that B or G is like RIGHT ARROW, and it should move to the next sequence.  Let's try this and see how it feels in practice over a longer period of time.