#752: prefix comments
Comments can now be added ( in parens ) at the beginning of a call.  In the file, they will show up the same way.  Internally, the parens are translated to "insert a comment\ncomment" and for display the curly braces are changed back to parens.  Parens are what CSDS Online uses.