#752: add Time-in-tip display to SD page
When we are NOT in Unlock/Edit mode (e.g. when we first come into the SD tab), we are in Playback mode.  In this mode, the first item in the Current Sequence pane is selected, and we'll have a little Time-in-tip timer below the sequence (where the resolve line usually is).  When in Unlock/Edit mode, the Time-in-tip goes away, and we see the Resolve line.  Standard keyboard shortcuts (in particular, "P" and period will work while in the SD tab.  So, while in the SD tab, we can see where we are in the tip, AND we can control the music player.