#752: commented out some older sdsequence code
I think this code is no longer in use (it's commented out almost everywhere).  Commented out the remaining pieces.  We might want to bring back the use of QLabels instead of QTableWidgetItem in the Current Sequence pane, so that the text can be styled, so leaving the code in there (just commented out).