#752: introduce highlighting tag
We can now stick a tag onto a sequence in a sequence file, that specifies which calls are to be highlighted.  Highlighted calls will be in RED.  The highlight tag ("#HIGHLIGHT=call1,call2#") only affects that one sequence.  Any number of calls can be highlighted.  It's a simple string compare, NOT a QRegularExpression, so "square thru" will match "square thru {1,2,3,4}".