#834: Sequence Titles
Each sequence can now have an optional title string.  This string will be drawn at the top of the sequence, in BOLD, Italic, Underlined, to delimit it from the calls.  The Sequence Title can be changed by clicking UNLOCK.  It will be saved in the file as "#TITLE=<string>#", with any "#" characters escaped to "%%%".