#844: playlist numbers don't stay within legal bounds
Using the CMD-SHIFT-UP key, it was possible to make the item number in a playlist less than 1.  Similarly, using the CMD-SHIFT-DOWN key, it was possible to make the item number in a playlist greater than the number of items on the playlist.  But wait, there's more!  It was also possible to use those two keys on NON-playlist items, and hilarity ensued.