#467: initial sd.ini-stye abbreviations
This is equivalent to what SD does with sd.ini (and in fact it uses the same file .ini file, copied Contents/MacOS).  But that's not as good as Taminations.  So, I'll check this in for now as a partial solution (and to record some minimal info on how the SD abbreviation system works), and we can do something more Taminations-like later.