#856: better abbreviations system
This works much better than the sd.ini thing.  Any token (word separated by whitespace or a bracket) will be replaced with a string from the abbrevs.txt file.  If such a file does not exist in the musicRootDir, a default one will be created that can be manually edited.  Restart SquareDesk to pick up a modified file.