#799: CPU time improvements for the idle cases
In many use cases, if the Tip Timer field is not changing, the CPU time is now dropped from ~5% to ~1.5 - 2%.  This is about a 2.5X reduction in CPU time for these use cases.  The Tip Timer field is unchanging when nothing is loaded, or when a singing call is loaded and playback is stopped.  When the tip timer is ticking, say in the middle of a tip, the CPU time will still be around 5%.  To stop the Tip Timer, click on it once.