#884: Wrong BPM for songs...
There were a bunch of songs that were generating the wrong BPM.  I factored out the BPM code, and did some tests by voting between various start and end points.  The winner was just to sample between 60 and 90 seconds into the song.  It worked in all but 1 of my 15 problematic cases.  That's probably as close as we'll get with this algorithm (miniBPM).