#806: Part 1 of improved beat detection
This test code uses an external program from the vamp distribution.  Disabled by default. If enabled, we convert entire audio to mono, LPF to 1500Hz, write out the audio as WAV, and call vamp to do beat/measure detection.  Text results are sucked back in and compressed to a short string.  It then decompresses to make sure that compression and decompression worked.  Future parts will snap In/Out buttons to beats/measures, and beat/measure detection results will be cached in Sqlite.