#806: Part 2 of improved beat detection
This commit adds the snapping code.  It's disabled right now, controlled by a #define called TEST_GRANULARITY.  But, it's plumbed all the way down.  There is still a UX pause the first time a button is clicked, but I'll work on that next.  Works really good so far!  Still requires Vamp be externally compiled, and the path to it is hard-coded right now.  But, this basically demonstrates the concept.