#806: Part 7 of improved beat detection
I moved the VAMP part of beat detection to be asynchronous.  The only synchronous part now is the writing of the mono WAV file.  It's also clever about running VAMP only if Snap is ON (either Nearest Beat or Nearest Measure).  It's even smart enough to call VAMP when you switch modes AFTER a song has been loaded.  This adds about 0.5sec to the load time, which makes is slightly slower.  But, the bulk of the time (VAMP) is about 1.6 seconds, and is asynchronous.