#821: Pan/EQ Volume Compensation
The Constant Power Pan control (sine-based) introduces a factor of 0.707 when the Pan control is centered.  And, EQ introduces a factor of about 0.767 when all the EQ controls are at center position.  To compensate for this, a preference now allows for boosting the volume for all songs by -10 to +10dB.  Normally, this would be set at 5.5dB (results in a factor of 1.0/(0.707 * 0.767)), for those who want to compensate, and are willing to take the risk of clipping when non-default PAN or EQ settings are used.