#902: Too many VAMP messages
OK, so I plumbed this all the way down to the QAudioDecoder.  If vamp can't be found, and beat detection fails (for this or other reasons), the -time_sec is passed back, and the caller (who has access to ui->actionDisabled will disable the beat detection.  No more messages after that, unless you try to turn it on again.