#903: loop back missed
In some cases, when playing at a non-100% speed, we can get into a situation where the playPosition moves past the loopEndpoint, AND we don't detect that.  The song then plays to the end, like there was no loop.  This commit does 2 things: if there are less than 100 frames of space available in the m_audioSink, we just skip that and come back 5ms later. And, if we sneak past the endLoop Point and play too many samples, we loop back, and skip that many samples.