#796: user loads music while cuesheet is unlocked
If the cuesheet is unlocked, the cuesheet selector is now DISABLED, so that the user can't switch to a different cuesheet.  Also, if the user double-clicks to load a new music file, the user will now be asked to Save, Don't Save, or Cancel.  Save will save and then complete the load. Don't Save will NOT save and then complete the load.  Cancel will not save AND will not load the new song.