#921: initial hookup of darkSearch box to searching
If you click on one of the types under Tracks, the darkSearch field will be set to filter and show just the tracks with that type.  It's a bit weird right now, because we don't have a darkSongTable yet, so the sorting actually is only visible on the Music tab, and not yet on the DarkMode tab.