#921: initial darkSongTable support
It now displays all the songs now, and we can scroll through the list.  Nothing is hooked up yet though, so can't see playlists, can't double-click on songs, doesn't reflect filters, and changes to header widths are not persistent. The new darkLoadMusicList() function is about twice as fast as the old loadMusicList().