#921: dynamically assign Cmd-S and Shift-Cmd-S
Because there are now separate menu items for saving Playlists, Cuesheets, and SD Sequences, we now move around the Cmd-S and Shift-Cmd-S shortcuts, based on what tab the user is on.  Note: can't currently SAVE something, unless you're on that tab.  We might want to change the isEnabled for these menu items to always enabled, if this turns out to be too unexpected (but, it's essentially the way the single File > Cmd-S and File > Shift-Cmd-S menu items worked before, when there was just the FIle ones and they changed names dynamically).