#921: initial layout scaling (phase 1)
When the window is resized, the layout needs to breathe accordingly.  This can be remarkably tricky to do, while maintaining the alignment of all the widgets.  This is Phase 1 of that effort.  The lower half now has a movable separator (not persistent yet), in preparation for adding some playlist support, and making the songTable scale in X.  The upper left quadrant widgets are all in vertical/horizontal layouts now, in preparation for auto-scaling the waveform display in X sometime soon.