#921: Splitter works better now
The button to hide/show the playlists is currently now hidden (it was challenging to get it in there, so I've just hidden it for now, in case I want to bring it back later).
This is the basic layout for playlists now, but I'm thinking it can also be used for search results (e.g. Patter, Singers, Rounds, etc.).  More work is needed to understand how to connect the search box, double clicking, and tempo/pitch controls to these new "palette" buckets (they are similar to rekordbox).