#921: plumbing for playlists
New icon for palette slots.  I got rid of the QToolButton on the right, because I couldn't style it correctly.  The selections are now mutually exclusive between palettes and the songTables, in preparation for allowing ENTER to work (will consistently load the selected/highlighted song, regardless of which palette or songTable).  Also, prototype context menu on the palette slot titles, in preparation for some cool stuff soon.