#921: Context menu on TreeWidget playlist items
RIGHT-CLICK on a Playlist in the TreeWidget, and you can now show that playlist in one of the three Palette slots.  The context menu in the Palette slot header now tells you which file you can Edit or Reload.  The Pallete slots now also have an additional context menu item to Remove from Palette.  At startup, the last-loaded (via Playlist > Load) playlist will also be shown in Palette slot #1.  Finally, if there is no playlist in a slot, its title will be "Untitled Playlist", in anticipation of being able to add items to it soon.