#921: correctly load xtras, et.al.
Fixes a bug whereby the palette slot was not getting filled with the filtered song list at startup time (worked fine at double-click time).  Problem was that double-click is an extension of single-click, which filtered the darkSongTable.  At startup, this didn't happen, since there were no filters yet.  Fixed it so that track filtering by song type is correct at all times (including at startup), regardless of the state of the darkSongTable.