#928: restore non-DARKMODE initial load of playlist
I had commented out some lines which should not have been commented out.  While I was fixing that, I provided a SMALL amount of forward compatibility with DARKMODE.  So, the latest top-of-tree should be able to see and use playlists set by 1.0.5 .  However, the reverse is not true.  Once you start using hierarchical playlists, you gotta stick to the top-of-tree version and later.