#921: allow separate editing of Pitch/Tempo in playlists AND darkSongTable
If you want to edit the pitch/tempo of a playlist item, make sure that the playlist item is showing in one of the palette slots.  Then, double-click to load that item.  Change the pitch/tempo sliders, and just that one playlist item will be changed.  If you double-click from the songTable or darkSongTable, then changing pitch/tempo sliders will change it in the songTables, in any matching items in a Track Filter list, and in the sqlite3 database (the way it works today) -- in this case, the playlist items will NOT change.