#921: GUI affordance for "this is loaded"
If a song is loaded from a Playlist or from the darkSongTable, a pencil icon will appear before the title that now has editable pitch and tempo, and the text will be bold/italic.  This will not happen to the darkSongTable, to preserve compatibility with the old songTable.  I don't know if this will make sense or not, I'm trying it.  In any case, the code is there to be able to make highlighting changes to QTableWidgetItems, and to remove that highlighting at the right time.