#921: fix spurious build stop when MacOS is a file instead of a dir
This fixes a long-standing problem with the test123.pro file, wherein a brand new Qt version forces us to create a new build directory, and that build stops because Contents/MacOS is a file and not a directory/folder.  Now it should create the Contents/MacOS directory correctly when it's not there.