#921: show playlist items in canonical format
Previous to this commit, items in a playlist were shown as their original filenames, which could be a mixture of formats (e.g. "RIV123-foo", "foo-RIV123", "RIV 123 - foo", etc).  With this commit, playlist items are always seen in a canonical format, e.g. "RIV123 - foo", regardless of what format the original filename is in.  If a filename can't be figured out by us, it just shows up as the original filename.  This is something of an editorial decision, because it favors one particular format over others, for consistency sake.  This is a simplifier for now, until Playlists and Track Filters have separate columns for label and title (like the songTables do).