#921: Playlist double click should use shortTitle
Prior to this commit, double clicking on a playlist item would set the main title to be the filename, e.g. "RIV 123 - foo".  This is different from double clicking on an item in the songTable/darkSongTable, which sets the main title to something like "foo".  With this commit, double clicking on playlist items now works the same as double clicking on songTable items.