#1015: part one of patter segmentation
If a patter recording is loaded, a "Segment" button will appear.  Pressing that button will open a dialog asking if you're sure.  If you click YES, it will begin segmentation.  When complete, it will process the results file in audiodecoder.  Nothing happens after that, at this point -- this is part 1 of 2.  Commenting IN L9691 enables the Segment button for testing.