#1047: change definition of song "played"
A song now needs to play for 15 seconds in a session, before it is marked as "played".   If a song plays over a session boundary, it might be marked "played" in the earlier session (if it had more than 15 seconds of playback there), and also in the later session (if it had more than 15 seconds of playback there).