#1047: green strikethrough for played songs
When a song has played in the current session for >=15 seconds, we'll overlay a green strikethrough, both in the darkSongTable and in the palette slots, to indicate that the song has been "played".  Use "Clear Recent" to reset the fence to "now()", thereby removing the strikethroughs.  Use "Don't Save Plays" to disable strikethrough (and, I think, recording plays to the current session).  I think we might want a user preference to View/Not View the strikethroughs.