#1060: update Xcode
QMAKE_MAC_SDK is now macosx14.4 .  I also put some comments in to make it easier to upgrade next time.  I also changed the .pro file so that it would correctly make the MacOS folder before trying to stick the sd_calls.dat file there (which makes a directory called "MacOS" which is NOT a directory.