#1065: added dependencies for vamp
vamp-simple-host should be self-contained now to the Contents/MacOS folder.  It previously was reaching out to /usr/local/opt, requiring libsnd and ITS dependencies to be on the system...  that should be fixed now, and vamp should run on both X86_64 and M1 systems (with Rosetta), even if you haven't installed libsndfile with Homebrew.  I included the original dylibs from libsndfile from Homebrew, and a doit script that will fix all the library paths so that files are localized to Contents/MacOS.