The Cuesheet "Save as" menu item is now enabled when the patter.template.html is being edited.
The intent is as long as no lyrics or cuesheets are loaded, then when a patter doesn't have a following singer (so the "peek thing" works) one can add cuesheets to a patter.  As soon as a real cuesheet is loaded this behaviour stops,  cuesheets remain.   When editing the patter.template.html, though, the "Save as" re-loads the cuesheet which set cueSheetLoaded true so it could only be done once.   To prevent this, the "Save as" action saves and restores the value of cueSheetLoaded around the call to saveLyricsAs.