#1087: pathnames with double quotes now write out to CSV correctly
Double quotes need to be written as doubleQuote doubleQuote to be valid CSV.  If the simple CSV parser encounters a naked double quote in a CSV file, yes the CSV file is malformed, but squareDesk will crash.  In this case the only thing you can do is to delete the offending line from the CSV file.