The right button menu items now work in lightmode. The current playlist, shown at the bottom left corner, is now visible being black (previously it was almost the same as the background colour). To move a song down in the playlist, Command+Down doesn't work but Command+Option+Down does;  I don't know why.
Changes summarized:
In mainwindow.cpp:
Variable darkTextColor was being used before set, and was inappropriate for lightmode.
Debug calls (commented out) in GlobalEventFilter::eventFilter only confirmed that it wasn't called with Command+DownArwow.

In MainWindow::on_songTable_itemSelectionChanged the "bad SelectRow 2" message seemed useless and the call after did nothint so they were commented out.

MainWindow::on_songTable_customContextMenuRequested was basically empty because DARKMODE was defined so the ifdef removed and various SLOT parameters changed to their new names.

In playlists.cpp:
When in lightmode the "playlists/" part of the playlist name was removed.
A lot of code in MainWindow::PlaylistItemsRemove was commented out assuming dark mode.
I uncommented it if in lightnode.