#1092: don't overwrite an unsaved palette slot
If all the slots are full, normally the last slot would be overwritten on double-click in the tree widget.  BUT, if that slot is not really a playlist yet, because it's an UNSAVED playlist with items in the list, we do NOT want to overwrite that.  In this case, we now just put up a warning message, and let the user deal with it.