#1051: save playlist when pitch/tempo changes affect it
Pitch/tempo affect a playlist item if (and only if) an item from that playlist has been loaded.  Also, a subtle clue is given when a playlist item has been loaded:  the row number and pitch/tempo columns are bold/italicized.  Also, visibility of pitch and tempo columns in the playlists now follow the View > Columns > {Pitch, Tempo} settings in the View menu.