#1114: SquareDesk Help links to online SquareDesk Wiki
And, squaredesk.pdf is removed from the build process, and it's no longer copied to the reference directory in the music directory.  It should be easier to make changes to the doc, now that it updates immediately for everybody.  The tradeoff is that it's online, so internet access is required to read it now.