#1132: eliminate crashes at Light/Dark switch time
Moved cleanLockFile up, so it would clear right away in the destructor.  We now do NOT try to acquire the mutexThreadRunning, because that could hang if SD was hung, preventing shutdown.  Now we check every time iofull is dereferenced, because sometimes it could be null (if now if it was, it's logged in ALL places that might happen).  Added code to do a restart Light/Dark mode test, and ran that test more than 100 times (now, no crashes, and it works every time).