#1137: fixes problem with non-standard folder names not recognized as that category of song
There are a bunch of places in the code that look for "singing" in the filename, even though singing calls might be in a folder called "singers", where "singers" was in the Preferences > Music tab as an allowed type for singing calls.

So, if the folder name isn't "singing", it isn't recognized as a singing call, and the auto-switch-to-Cuesheet-tab feature doesn't work.

It turns out that this problem existed in a number of places, including playlists and the "peek" feature.  This commit fixes all of those places (I think!).