#102: determine sample rate of ORIGINAL MP3 file
When we read in files, we tell Qt to auto-resample everything to 44.1KHz, so that all our pipelines work the same way (we don't have to regenerate anything at load time).  But, setting a LOOPSTART ID3V2 tag will require setting in SAMPLES, so we need to convert our 44.1KHz loop start to a 48KHz loop start for about 6% of the files (which on my system are 48KHz.