#1270: undo previous commit
If we put the musicDir itself on the musicrootwatcher's watch list, it will trigger when we make a new SD sequence, because that writes a file.  That will put the focus back into the search field on the dark music tab, which we do not want.  We must fix this better later, maybe screen for being on the SD page?