#1268: Importing playlists from Apple Music
It used to be that M3U was the only export format that Apple Music did, that SquareDesk could import.  M3U support went away long ago, though.  So, this is a way, hopefully better, for importing from Apple Music.  It does require that Apple Music be running when SquareDesk is started.