#1287: preFiltering for performance
It's now way faster to click on a TreeWidget item, because I filter out almost everything BEFORE it goes into the darkSongTable.  The UI is now consistent: 1) clicking on a TreeWIdgetItem filters and then sticks into the darkSongTable, 2) the darkSearch field then further filters whatever is in the darkSongTable.  This seems like the right UX to me, and it's way faster than sticking everything into the darkSongTable and using the darkSearch field to filter it down.