#1309: Can't escape from cuesheet Unlock
If a song has zero cuesheet matches, and this is often true for non-square-dance tunes, then clicking on Unlock for Editing gets rid of the buttons, and you can't escape out of it.  Furthermore, it tantalized you with a lyrics template that was maybe the wrong template, but then it told you to save it at app exit time. Now it says "no cuesheets", does NOT show any template, and if you click Unlock, it suggests what you should do.