#1379: last_cuesheet should be RELATIVE, not ABSOLUTE
Otherwise, it will lose cuesheet associations when the music directory is moved somewhere else.  I think I fixed these, so that they are still absolute in the DB, but they are always validated and converted to the current music root when referenced (e.g. by Reveal or by setting the default cuesheetComboBox item).  So, if you move the musicRoot, over time the cuesheet paths will be converted to the new musicRootPath.  They are still absolute though.